// packages/chart-lib/lightweight-charts/plugins/bands-indicator.d.ts
// Generated by dts-bundle-generator v9.5.1

import { CanvasRenderingTarget2D } from 'fancy-canvas';
import { AutoscaleInfo, Coordinate, DataChangedScope, IChartApi, IPrimitivePaneRenderer, IPrimitivePaneView, ISeriesApi, ISeriesPrimitive, Logical, SeriesAttachedParameter, SeriesDataItemTypeMap, SeriesOptionsMap, SeriesType, Time } from 'lightweight-charts';

declare abstract class PluginBase implements ISeriesPrimitive<Time> {
	private _chart;
	private _series;
	protected dataUpdated?(scope: DataChangedScope): void;
	protected requestUpdate(): void;
	private _requestUpdate?;
	attached({ chart, series, requestUpdate }: SeriesAttachedParameter<Time>): void;
	detached(): void;
	get chart(): IChartApi;
	get series(): ISeriesApi<keyof SeriesOptionsMap>;
	private _fireDataUpdated;
}
export type SearchDirection = "left" | "right";
declare class ClosestTimeIndexFinder<T extends {
	time: number;
}> {
	private numbers;
	private cache;
	constructor(sortedNumbers: T[]);
	findClosestIndex(target: number, direction: SearchDirection): number;
	private _performSearch;
}
export interface UpperLowerData {
	upper: number;
	lower: number;
}
declare class UpperLowerInRange<T extends UpperLowerData> {
	private _arr;
	private _chunkSize;
	private _cache;
	constructor(arr: T[], chunkSize?: number);
	getMinMax(startIndex: number, endIndex: number): UpperLowerData;
	private _check;
}
export interface BandRendererData {
	x: Coordinate | number;
	upper: Coordinate | number;
	lower: Coordinate | number;
}
declare class BandsIndicatorPaneRenderer implements IPrimitivePaneRenderer {
	_viewData: BandViewData;
	constructor(data: BandViewData);
	draw(): void;
	drawBackground(target: CanvasRenderingTarget2D): void;
}
export interface BandViewData {
	data: BandRendererData[];
	options: Required<BandsIndicatorOptions>;
}
declare class BandsIndicatorPaneView implements IPrimitivePaneView {
	_source: BandsIndicator;
	_data: BandViewData;
	constructor(source: BandsIndicator);
	update(): void;
	renderer(): BandsIndicatorPaneRenderer;
}
export interface BandData {
	time: Time;
	upper: number;
	lower: number;
}
export interface BandsIndicatorOptions {
	lineColor?: string;
	fillColor?: string;
	lineWidth?: number;
}
export declare class BandsIndicator extends PluginBase implements ISeriesPrimitive<Time> {
	_paneViews: BandsIndicatorPaneView[];
	_seriesData: SeriesDataItemTypeMap[SeriesType][];
	_bandsData: BandData[];
	_options: Required<BandsIndicatorOptions>;
	_timeIndices: ClosestTimeIndexFinder<{
		time: number;
	}>;
	_upperLower: UpperLowerInRange<BandData>;
	constructor(options?: BandsIndicatorOptions);
	updateAllViews(): void;
	paneViews(): BandsIndicatorPaneView[];
	attached(p: SeriesAttachedParameter<Time>): void;
	dataUpdated(scope: DataChangedScope): void;
	_minValue: number;
	_maxValue: number;
	calculateBands(): void;
	autoscaleInfo(startTimePoint: Logical, endTimePoint: Logical): AutoscaleInfo;
}

export {};